+++ Sadie.Game.Rooms/IRoomData.cs
@@ -19,6 +19,6 @@ public interface IRoomData
     IRoomUserRepository UserRepository { get; }
     IRoomSettings Settings { get; }
     List<RoomChatMessage> ChatMessages { get; }
-    public List<int> PlayersWithRights { get; }
+    public List<long> PlayersWithRights { get; }
     public IRoomFurnitureItemRepository FurnitureItemRepository { get; }
 }
